Move getSeries logic to the JaiaH5FileSet class
