+++ src/web/jdv/server/jaialogs.py
@@ -37,43 +37,6 @@ def set_directory(directory):
     logging.info(f'Log directory: {directory}')
     os.makedirs(directory, exist_ok=True)
 
-# Path descriptions
-
-try:
-    path_descriptions = json.load(open('jaiabot_paths.json'))
-except FileNotFoundError:
-    path_descriptions = {}
-
-
-def jaia_get_description(path):
-    for description in path_descriptions:
-        if 'path' in description:
-            if description['path'] == path:
-                return description
-        
-        if 'path_regex' in description:
-            if re.match(description['path_regex'], path):
-                return description
-    
-    return None
-
-def get_title_from_path(path: str):
-    components = path.split('/')
-    if len(components) < 2:
-        logging.warning(f'Not enough components in path: {path}')
-        return ''
-
-    components = components[1:]
-
-    message_type_components = components[0].split('.')
-
-    if len(message_type_components) < 1:
-        logging.warning(f'Invalid path: {path}')
-        return ''
-
-    components[0] = components[0].split('.')[-1]
-    return '/'.join(components)
-
 # Data fetch functions
 
 UTIME_PATH = 'goby::health::report/goby.middleware.protobuf.VehicleHealth/_utime_'
@@ -123,53 +86,16 @@ def get_fields(log_names: List[str], root_path='/'):
     return h5_files.fields(root_path=root_path)
 
 
-def get_series(log_names, paths):
+def get_series(log_names: List[str], paths: List[str]):
     '''Get a series'''
 
     if log_names is None or paths is None:
         return []
 
-    series_list = []
+    h5_paths = [f'{LOG_DIR}/{name}.h5' for name in log_names]
+    h5_files = JaiaH5FileSet(h5_paths, shouldConvertGoby=True)
 
-    log_names = log_names.split(',')
-    paths = [path.lstrip('/') for path in paths.split(',')]
-
-    # Open all our logs
-    logs = [h5py.File(log_name) for log_name in log_names]
-
-    # Get the series from the logs
-    for path in paths:
-        series_description = jaia_get_description(path) or {}
-        invalid_values = set(series_description.get('invalid_values', []))
-
-        series = Series()
-
-        for log in logs:
-            try:
-                series += Series(log=log, path=path, scheme=1, invalid_values=invalid_values)
-            except KeyError as e:
-                logging.warn(e)
-                continue
-
-        series.sort()
-
-        title = get_title_from_path(path)
-        y_axis_title = title
-        units = series_description.get('units')
-
-        if units is not None:
-            y_axis_title += f'\n({units})'
-
-        series_list.append({
-            'path': path,
-            'title': title,
-            'y_axis_title': y_axis_title,
-            '_utime_': series.utime,
-            'series_y': series.y_values,
-            'hovertext': series.hovertext
-        })
-
-    return series_list
+    return h5_files.getSeries(paths)
 
 
 def get_map(log_names: List[str]):
